+++ src/neural_network/layer/Convolution1D.cpp
@@ -57,7 +57,7 @@ void Convolution1D::insertBackOutputForNeuron(int neuronNumber, const std::vecto
 {
     neuronNumber = neuronNumber % this->getNumberOfNeurons() / this->numberOfFilters;
     const int beginIndex = neuronNumber * this->shapeOfInput[1];
-    for (int e = 0; e < error.size(); ++e)
+    for (int e = 0; e < (int)error.size(); ++e)
     {
         const int i = beginIndex + e;
         errors[i] += error[e];
